--update: Update/Change the pages name
