+++ lib/pages/filesPage.dart
@@ -1,7 +1,7 @@
 // ignore_for_file: must_be_immutable
 
 import 'package:flutter/material.dart';
-import 'package:lvlmindbeta/pages/homepage.dart';
+import 'package:lvlmindbeta/pages/homePage.dart';
 import 'package:lvlmindbeta/Models/profs.dart';
 import 'package:lvlmindbeta/Models/matiere.dart';
